+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added better diagnostic with `ariadne`.
+
+### Changed
+
+- Bumped `mml-lib@v0.3.2`.
+
 ## [0.2.1] - 2023-08-30
 
 ### Changed